macro get_columns_in_relation to respect the temp metadata generation schema config
